// apps/web/src/shared/api/generated/model/socialFeedSliceResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */
import type { SocialFeedResponse } from './socialFeedResponse';

/**
 * 피드 슬라이스 응답
 */
export interface SocialFeedSliceResponse {
  /** 다음 페이지 존재 여부 */
  hasNextPage: boolean;
  /** 피드 목록 */
  contents: SocialFeedResponse[];
}